+++ packages/stream_video_flutter/lib/src/call_participants/call_participant.dart
@@ -40,6 +40,7 @@ class StreamCallParticipant extends StatelessWidget {
     this.audioLevelIndicatorColor,
     this.enabledMicrophoneColor,
     this.disabledMicrophoneColor,
+    this.pausedVideoIndicatorColor,
     this.showConnectionQualityIndicator,
     this.connectionLevelActiveColor,
     this.connectionLevelInactiveColor,
@@ -94,6 +95,9 @@ class StreamCallParticipant extends StatelessWidget {
   /// The color of a disabled microphone icon.
   final Color? disabledMicrophoneColor;
 
+  /// The color of a paused video indicator.
+  final Color? pausedVideoIndicatorColor;
+
   /// Whether to show the connection quality indicator.
   final bool? showConnectionQualityIndicator;
 
@@ -140,6 +144,8 @@ class StreamCallParticipant extends StatelessWidget {
         this.enabledMicrophoneColor ?? theme.enabledMicrophoneColor;
     final disabledMicrophoneColor =
         this.disabledMicrophoneColor ?? theme.disabledMicrophoneColor;
+    final pausedVideoIndicatorColor =
+        this.pausedVideoIndicatorColor ?? theme.pausedVideoIndicatorColor;
     final showConnectionQualityIndicator =
         this.showConnectionQualityIndicator ??
             theme.showConnectionQualityIndicator;
@@ -238,6 +244,7 @@ class StreamCallParticipant extends StatelessWidget {
                           audioLevelIndicatorColor: audioLevelIndicatorColor,
                           disabledMicrophoneColor: disabledMicrophoneColor,
                           enabledMicrophoneColor: enabledMicrophoneColor,
+                          pausedVideoIndicatorColor: pausedVideoIndicatorColor,
                           participantLabelTextStyle: participantLabelTextStyle,
                         ),
                       ],